+++ src/main/java/com/bginfosys/dinghyracing/model/Crew.java
@@ -1,0 +1,47 @@
+/*
+ * Copyright 2022-2024 BG Information Systems Ltd
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License. 
+ */
+
+package com.bginfosys.dinghyracing.model;
+
+import com.bginfosys.dinghyracing.model.Competitor;
+
+public class Crew {
+
+private Competitor helm;
+	
+	private Competitor crew;
+
+	public Crew(Competitor helm, Competitor crew) {
+		this.helm = helm;
+		this.crew = crew;
+	}
+
+	public Competitor getHelm() {
+		return helm;
+	}
+
+	public void setHelm(Competitor helm) {
+		this.helm = helm;
+	}
+
+	public Competitor getCrew() {
+		return crew;
+	}
+
+	public void setCrew(Competitor crew) {
+		this.crew = crew;
+	}
+}
