+++ src/main/java/com/bginfosys/dinghyracing/web/controllers/CrewController.java
@@ -1,3 +1,19 @@
+/*
+ * Copyright 2022-2024 BG Information Systems Ltd
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License. 
+ */
+
 package com.bginfosys.dinghyracing.web.controllers;
 
 import java.net.URI;
@@ -23,7 +39,7 @@ import org.springframework.web.bind.annotation.RestController;
 
 import com.bginfosys.dinghyracing.model.Dinghy;
 import com.bginfosys.dinghyracing.persistence.CompetitorRepository;
-import com.bginfosys.dinghyracing.web.dto.CrewDTO;
+import com.bginfosys.dinghyracing.model.Crew;
 
 @RestController
 @RequestMapping("/dinghyracing/api/crews")
@@ -49,9 +65,7 @@ public class CrewController {
 	}
 	
 	@GetMapping(path= "/search/findCrewsByDinghy")
-//	public ResponseEntity<Set<CrewDTO>> findCrewsByDinghy(@RequestParam(name = "dinghy") String dinghyUri) {
 	public ResponseEntity<Object> findCrewsByDinghy(@RequestParam(name = "dinghy") String dinghyUri) {
-//		ResponseEntity<Set<CrewDTO>> responseEntity;
 		ResponseEntity<Object> responseEntity;
 		
 		// get dinghy matching supplied uri
@@ -65,16 +79,13 @@ public class CrewController {
 					return result;
 				}, dinghy.getId());
 		
-		Set<CrewDTO> crews = new HashSet<CrewDTO>();
+		Set<Crew> crews = new HashSet<Crew>();
 		for (Long[] result : results) {			
-			crews.add(new CrewDTO(competitorRepository.findById(result[0]).orElse(null), competitorRepository.findById(result[1]).orElse(null)));
+			crews.add(new Crew(competitorRepository.findById(result[0]).orElse(null), competitorRepository.findById(result[1]).orElse(null)));
 		}
 		
-		CollectionModel<CrewDTO> resource = CollectionModel.of(crews);
+		CollectionModel<Crew> resource = CollectionModel.of(crews);
 		
-//		responseEntity = ResponseEntity.ok()
-//			.header("Content-Type", "application/hal+json")
-//			.body(crews);
 		responseEntity = ResponseEntity.ok()
 			.header("Content-Type", "application/hal+json")
 			.body(resource);
